backend/requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
certifi==2022.12.7
# cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==39.0.0
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.8.0
Django==4.1.6
django-cors-headers==3.13.0
django-countries==7.5.1
django-stubs==1.14.0
django-stubs-ext==0.7.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mock==4.0.3
mypy==0.991
mypy-extensions==0.4.3
netaddr==0.8.0
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
yookassa==2.3.4
zope.interface==5.5.2
Cmake